Refaktorisi vracanje gresaka u kontrolerima (#68)

+++ src/main/java/rs/raf/demo/controllers/FakturaRestController.java
@@ -15,6 +15,7 @@ import rs.raf.demo.utils.ApiUtil;
 
 import rs.raf.demo.specifications.RacunSpecificationsBuilder;
 
+import javax.persistence.EntityNotFoundException;
 import javax.validation.Valid;
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
@@ -22,7 +23,6 @@ import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.Optional;
-import java.util.Map;
 
 @CrossOrigin
 @RestController
@@ -39,32 +39,17 @@ public class FakturaRestController {
 
     @GetMapping(value = "/ulazneFakture", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<?> getUlazneFakture() {
-        List<Faktura> ulazneFakture = fakturaService.findUlazneFakture();
-        if(ulazneFakture.isEmpty()){
-            return ResponseEntity.notFound().build();
-        }else{
-            return ResponseEntity.ok(ulazneFakture);
-        }
+        return ResponseEntity.ok(fakturaService.findUlazneFakture());
     }
 
     @GetMapping(value = "/izlazneFakture", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<?> getIzlazneFakture() {
-        List<Faktura> izlazneFakture = fakturaService.findIzlazneFakture();
-        if(izlazneFakture.isEmpty()){
-            return ResponseEntity.notFound().build();
-        }else{
-            return ResponseEntity.ok(izlazneFakture);
-        }
+        return ResponseEntity.ok(fakturaService.findIzlazneFakture());
     }
 
     @GetMapping(value = "/sume", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<?> getSume(@RequestParam String tipFakture){
-        Map<String, Double> sume = fakturaService.getSume(tipFakture);
-        if(!sume.isEmpty()) {
-            return ResponseEntity.ok(sume);
-        } else {
-            return ResponseEntity.notFound().build();
-        }
+        return ResponseEntity.ok(fakturaService.getSume(tipFakture));
     }
 
     @GetMapping(value = "/all", produces = MediaType.APPLICATION_JSON_VALUE)
@@ -88,18 +73,9 @@ public class FakturaRestController {
 
         Specification<Faktura> spec = builder.build();
 
-        try{
-            List<Faktura> result = fakturaService.findAll(spec);
+        List<Faktura> result = fakturaService.findAll(spec);
 
-            if(result.isEmpty()) {
-                return ResponseEntity.notFound().build();
-            }
-
-            return ResponseEntity.ok(result);
-        }
-        catch (RuntimeException e){
-            return ResponseEntity.internalServerError().body(e.getMessage());
-        }
+        return ResponseEntity.ok(result);
     }
 
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
@@ -112,14 +88,19 @@ public class FakturaRestController {
         Optional<Faktura> optionalFaktura = fakturaService.findById(faktura.getDokumentId());
         if(optionalFaktura.isPresent()) {
             return ResponseEntity.ok(fakturaService.save(faktura));
-        } else {
-            return ResponseEntity.notFound().build();
         }
+
+        throw new EntityNotFoundException();
     }
 
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<?> deleteFaktura(@PathVariable("id") Long id){
-        fakturaService.deleteById(id);
-        return ResponseEntity.ok().build();
+        Optional<Faktura> optionalFaktura = fakturaService.findById(id);
+        if (optionalFaktura.isPresent()){
+            fakturaService.deleteById(id);
+            return ResponseEntity.noContent().build();
+        }
+
+        throw new EntityNotFoundException();
     }
 }
